:truck::ok_hand: Move camel_case function to string_utils module.
